+++ src/config/keybinds.rs
@@ -334,6 +334,8 @@ pub struct Keybinds {
     pub rename_pane: ActionKeybinds,
     pub edit_scrollback: ActionKeybinds,
     pub copy_mode: ActionKeybinds,
+    pub jump_prompt_up: ActionKeybinds,
+    pub jump_prompt_down: ActionKeybinds,
     pub focus_pane_left: ActionKeybinds,
     pub focus_pane_down: ActionKeybinds,
     pub focus_pane_up: ActionKeybinds,
@@ -496,6 +498,8 @@ impl Config {
             rename_pane: empty_action!(),
             edit_scrollback: empty_action!(),
             copy_mode: empty_action!(),
+            jump_prompt_up: empty_action!(),
+            jump_prompt_down: empty_action!(),
             focus_pane_left: empty_action!(),
             focus_pane_down: empty_action!(),
             focus_pane_up: empty_action!(),
@@ -637,6 +641,8 @@ impl Config {
             apply_action!(keybinds.rename_pane, rename_pane, source);
             apply_action!(keybinds.edit_scrollback, edit_scrollback, source);
             apply_action!(keybinds.copy_mode, copy_mode, source);
+            apply_action!(keybinds.jump_prompt_up, jump_prompt_up, source);
+            apply_action!(keybinds.jump_prompt_down, jump_prompt_down, source);
             apply_action!(keybinds.focus_pane_left, focus_pane_left, source);
             apply_action!(keybinds.focus_pane_down, focus_pane_down, source);
             apply_action!(keybinds.focus_pane_up, focus_pane_up, source);